+++ ZOHO doc/src/com/controller/NewFileController.java
@@ -7,6 +7,7 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.PrintWriter;
+import java.util.Arrays;
 import java.util.LinkedHashMap;
 
 import javax.servlet.ServletConfig;
@@ -15,6 +16,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.json.simple.JSONObject;
 
@@ -34,6 +36,8 @@ public class NewFileController extends HttpServlet {
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
+		HttpSession session = request.getSession();
+		String sessionUser = (String) session.getAttribute("user");
 		BufferedReader reader = new BufferedReader(new InputStreamReader(request.getInputStream()));
 		String fileName = request.getParameter("filename");
 		String mode = request.getParameter("mode");
@@ -42,25 +46,40 @@ public class NewFileController extends HttpServlet {
 		JSONObject jsonObject = new JSONObject();
 		String successState = "false";
 		File file = new File(defaultLocation + location);
-		if (!file.exists() && mode.equals("new")) {
-			file.createNewFile();
-			FileWriter fw = new FileWriter(file);
-			fw.write(text);
-			fw.close();
-			ClientsInfoDao.insertFile(location);
-			successState = "true";
-			AddWordsTask.getFileList().addFileName(location);
-		} else if (file.exists() && mode.equals("edit")) {
-			LinkedHashMap<Integer, String>[] twoLists = Utilities.getEditedWords(
-					Utilities.stringBuilder(new BufferedReader(new FileReader(file))), text);
-			AddWordsTask.getEditList().addFileName(ClientsInfoDao.getFileId(location)+"+"+location, twoLists);
-			file.delete();
-			file.createNewFile();
-			FileWriter fw = new FileWriter(file);
-			fw.write(text);
-			fw.close();
-			ClientsInfoDao.insertFile(location);
-			successState = "true";
+		boolean flag = false;
+		String checkLocation = location.substring(0, location.indexOf('/'));
+		if (checkLocation.equals(sessionUser)) {
+			flag = true;
+		} else {
+			long fileId = ClientsInfoDao.getFileId(request.getParameter("location"));
+			String privilegeInfo = ClientsInfoDao.checkLocation(fileId, sessionUser);
+			if (privilegeInfo.substring(0, privilegeInfo.indexOf('+')).equals("write")) {
+				flag = true;
+			}
+		}
+		System.out.println(file);
+		if (flag) {
+			if (!file.exists() && mode.equals("new")) {
+				file.createNewFile();
+				FileWriter fw = new FileWriter(file);
+				fw.write(text);
+				fw.close();
+				ClientsInfoDao.insertFile(location);
+				successState = "true";
+				AddWordsTask.getFileList().addFileName(location);
+			} else if (file.exists() && mode.equals("edit")) {
+				LinkedHashMap<Integer, String>[] twoLists = Utilities
+						.getEditedWords(Utilities.stringBuilder(new BufferedReader(new FileReader(file))), text);
+				System.out.println(Arrays.toString(twoLists));
+				AddWordsTask.getEditList().addFileName(ClientsInfoDao.getFileId(location) + "+" + location, twoLists);
+				file.delete();
+				file.createNewFile();
+				FileWriter fw = new FileWriter(file);
+				fw.write(text);
+				fw.close();
+				ClientsInfoDao.insertFile(location);
+				successState = "true";
+			}
 		}
 		jsonObject.put("success", successState);
 		response.setContentType("application/json");
